backend/requirements.txt
# Document parsing
llama-parse==0.4.7
pypdf==4.3.1
python-docx==1.1.2

# NLP — local
spacy==3.7.4
python-Levenshtein==0.23.0

# HuggingFace models
transformers==4.41.2
torch==2.3.0
sentence-transformers==3.0.1

# LLM APIs
anthropic==0.29.0
openai==1.35.0

# Vector database
pinecone-client==3.2.2

# Web framework + async
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
sse-starlette==2.1.0

# Async task queue
celery==5.4.0
redis==5.0.6

# Database
sqlalchemy==2.0.31
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Google APIs (Gmail)
google-api-python-client==2.134.0
google-auth-oauthlib==1.2.0

# Config + validation
pydantic==2.7.4
pydantic-settings==2.3.3
email-validator==2.2.0
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4

# Utilities
numpy==1.26.4
python-multipart==0.0.9
